refactor(web): split EventList into a view and a legacy container
